Test that trail has a direction and a path
