+++ pokemon.lua
@@ -2,6 +2,7 @@ module ( "pokemon", package.seeall )
 require "a-star"
 require "strlib"
 require "tile"
+require "trail"
 serpent = require "serpent"
 local inputbox = require "inputbox"
 scriptpath = debug.getinfo(1, "S").source:sub(2):match("^.*\\")
@@ -366,35 +367,6 @@ return false
 else
 return true
 end
-end
-
-function direction(x, y, destx, desty)
-local s = ""
-if y > desty then
-s = y-desty .. " up"
-elseif y < desty then
-s = desty-y .. " down"
-end
-if x > destx then
-s = s .. " " .. x-destx .. " left"
-elseif x < destx then
-s = s .. " " .. destx-x .. " right"
-end
-return s
-end
-
-function only_direction(x, y, destx, desty)
-local s = ""
-if y > desty then
-return "up"
-elseif y < desty then
-return "down"
-elseif x > destx then
-return "left"
-elseif x < destx then
-return "right"
-end
-return s
 end
 
 -- Read current and around tiles
